refactor: extract saves and gameLoop routers (PR-14, Phase 1)
Final pure-move PR of the Phase 1 server-routes refactor. Extracts the
saves and game-loop endpoints into feature routers, leaving routes.ts thin.

- server/routes/saves.ts: 5 endpoints (GET/POST /api/saves, GET/POST/DELETE
  by :saveId incl. the ~535-line restore handler). validateSnapshotCollections
  moves from the registerRoutes closure to module scope here (its only caller).
- server/routes/gameLoop.ts: POST /api/game/:gameId/actions, POST
  /api/advance-week, POST /api/select-actions (orphan, no auth — TODO-flagged
  for phase-2 deletion, behavior unchanged).

Routers mounted at the prior position of each router's first endpoint,
preserving global registration order and in-router relative order. Handlers
moved verbatim; routes.ts drops to ~121 lines with registerRoutes signature
unchanged.
